+++ components/bridge/rpc/api/src/main/java/com/flipperdevices/bridge/rpc/api/FlipperStorageApi.kt
@@ -2,4 +2,6 @@ package com.flipperdevices.bridge.rpc.api
 
 interface FlipperStorageApi {
     suspend fun mkdirs(path: String)
+
+    suspend fun delete(path: String, recursive: Boolean = false)
 }