+++ components/faphub/category/impl/src/main/java/com/flipperdevices/faphub/category/impl/api/FapHubCategoryApiImpl.kt
@@ -11,6 +11,7 @@ import com.flipperdevices.faphub.category.api.FapHubCategoryApi
 import com.flipperdevices.faphub.category.impl.composable.ComposableFapHubCategory
 import com.flipperdevices.faphub.dao.api.model.FapCategory
 import com.flipperdevices.faphub.fapscreen.api.FapScreenApi
+import com.flipperdevices.faphub.installation.button.api.FapButtonSize
 import com.flipperdevices.faphub.installation.button.api.FapInstallationUIApi
 import com.flipperdevices.faphub.installation.button.api.toFapButtonConfig
 import com.flipperdevices.faphub.search.api.FapHubSearchEntryApi
@@ -51,11 +52,11 @@ class FapHubCategoryApiImpl @Inject constructor(
                 onOpenFapItem = {
                     navController.navigate(fapScreenApi.getFapScreen(it.id))
                 },
-                installationButton = { fapItem, modifier, fontSize ->
+                installationButton = { fapItem, modifier ->
                     fapInstallationUIApi.ComposableButton(
                         config = fapItem?.toFapButtonConfig(),
                         modifier = modifier,
-                        textSize = fontSize
+                        fapButtonSize = FapButtonSize.COMPACTED
                     )
                 }
             )