+++ components/faphub/dao/network/src/main/java/com/flipperdevices/faphub/dao/network/retrofit/model/detailed/RetrofitApplicationDetailed.kt
@@ -58,7 +58,7 @@ data class RetrofitApplicationDetailed(
             picUrl = picUrl,
             metaInformation = metaInformation,
             fapDeveloperInformation = fapDeveloperInformation,
-            applicationId = alias,
+            applicationAlias = alias,
             upToDateVersion = fapItemVersion
         )
     }