+++ components/faphub/installation/manifest/impl/src/main/java/com/flipperdevices/faphub/installation/manifest/impl/api/FapManifestApiImpl.kt
@@ -5,8 +5,11 @@ import com.flipperdevices.core.ktx.jre.launchWithLock
 import com.flipperdevices.core.ktx.jre.withLock
 import com.flipperdevices.core.log.LogTagProvider
 import com.flipperdevices.core.log.error
+import com.flipperdevices.core.log.info
+import com.flipperdevices.core.log.warn
 import com.flipperdevices.faphub.dao.api.FapVersionApi
 import com.flipperdevices.faphub.installation.manifest.api.FapManifestApi
+import com.flipperdevices.faphub.installation.manifest.impl.utils.FapManifestDeleter
 import com.flipperdevices.faphub.installation.manifest.impl.utils.FapManifestUploader
 import com.flipperdevices.faphub.installation.manifest.impl.utils.FapManifestsLoader
 import com.flipperdevices.faphub.installation.manifest.model.FapManifestItem
@@ -29,7 +32,8 @@ import javax.inject.Singleton
 class FapManifestApiImpl @Inject constructor(
     private val loader: FapManifestsLoader,
     private val manifestUploader: FapManifestUploader,
-    private val fapVersionApi: FapVersionApi
+    private val fapVersionApi: FapVersionApi,
+    private val manifestDeleter: FapManifestDeleter
 ) : FapManifestApi, LogTagProvider {
     override val TAG = "FapManifestApi"
 
@@ -58,6 +62,23 @@ class FapManifestApiImpl @Inject constructor(
             } else {
                 null
             }
+        }
+    }
+
+    override suspend fun remove(
+        applicationId: String
+    ) = withLock(mutex, "remove") {
+        val toRemoveManifest = fapManifestItemFlow.value?.find {
+            it.uid == applicationId
+        }
+        if (toRemoveManifest == null) {
+            warn { "Can't find manifest for $applicationId" }
+            return@withLock
+        }
+        info { "Delete for $applicationId $toRemoveManifest" }
+        manifestDeleter.delete(toRemoveManifest)
+        fapManifestItemFlow.update { manifests ->
+            manifests?.minus(toRemoveManifest)
         }
     }