+++ components/faphub/installation/manifest/impl/src/main/java/com/flipperdevices/faphub/installation/manifest/impl/utils/FapManifestUploader.kt
@@ -33,7 +33,7 @@ class FapManifestUploader @Inject constructor(
         val newManifestPath = File(
             FAP_MANIFESTS_FOLDER_ON_FLIPPER,
             "${fapManifestItem.applicationAlias}.$FAP_MANIFEST_EXTENSION"
-        ).absolutePath
+        ).path
         atomicMover.atomicMove(
             pathToFap to fapManifestItem.path,
             tmpManifestPath to newManifestPath
@@ -46,7 +46,7 @@ class FapManifestUploader @Inject constructor(
         val tmpFapPath = File(
             tmpFolderProvider.provideTmpFolder(),
             "tmp.fim"
-        ).absolutePath
+        ).path
         uploadTmpManifest(serviceApi.requestApi, fapManifestItem, tmpFapPath)
         info { "Finish tmp manifest upload, path is $tmpFapPath" }
         return tmpFapPath