+++ components/faphub/installation/stateprovider/impl/src/main/java/com/flipperdevices/faphub/installation/stateprovider/impl/api/FapInstallationStateManagerImpl.kt
@@ -65,12 +65,14 @@ class FapInstallationStateManagerImpl @Inject constructor(
             is FapActionRequest.Cancel -> FapState.Canceling
             is FapActionRequest.Install -> FapState.InstallationInProgress(queueState.float)
             is FapActionRequest.Update -> FapState.UpdatingInProgress(queueState.float)
+            is FapActionRequest.Delete -> FapState.Deleting
         }
 
         is FapQueueState.Pending -> when (queueState.request) {
             is FapActionRequest.Cancel -> FapState.Canceling
             is FapActionRequest.Install -> FapState.InstallationInProgress(0f)
             is FapActionRequest.Update -> FapState.UpdatingInProgress(0f)
+            is FapActionRequest.Delete -> FapState.Deleting
         }
 
         FapQueueState.NotFound,