+++ components/faphub/installedtab/impl/src/main/java/com/flipperdevices/faphub/installedtab/impl/api/FapInstalledApiImpl.kt
@@ -3,6 +3,7 @@ package com.flipperdevices.faphub.installedtab.impl.api
 import androidx.compose.runtime.Composable
 import com.flipperdevices.core.di.AppGraph
 import com.flipperdevices.faphub.dao.api.model.FapItemShort
+import com.flipperdevices.faphub.installation.button.api.FapButtonSize
 import com.flipperdevices.faphub.installation.button.api.FapInstallationUIApi
 import com.flipperdevices.faphub.installation.button.api.toFapButtonConfig
 import com.flipperdevices.faphub.installedtab.api.FapInstalledApi
@@ -18,11 +19,11 @@ class FapInstalledApiImpl @Inject constructor(
     override fun ComposableInstalledTab(onOpenFapItem: (FapItemShort) -> Unit) {
         ComposableInstalledTabScreen(
             onOpenFapItem = onOpenFapItem,
-            installationButton = { fapItem, modifier, fontSize ->
+            installationButton = { fapItem, modifier ->
                 fapInstallationUIApi.ComposableButton(
                     config = fapItem?.toFapButtonConfig(),
                     modifier = modifier,
-                    textSize = fontSize
+                    fapButtonSize = FapButtonSize.COMPACTED
                 )
             }
         )